Change onboarding to func
