+++ src/Screens/index.js
@@ -2,3 +2,4 @@ export CommonsList from './CommonsList';
 export Home from './Home';
 export Login from './Login';
 export NativeBridgeTests from './NativeBridgeTests';
+export Onboarding from './Onboarding';
